bell24h-main/client/oracle-deployment/ml-service/main.py
"""
Bell24h ML Service - SHAP/LIME Explainability
Deployed on Oracle Cloud ARM VM (1GB)
"""

from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
import pandas as pd
import numpy as np
import joblib
import redis
import json
import hashlib
from typing import Dict, Any, List
import os
from datetime import datetime, timedelta

# Initialize FastAPI app
app = FastAPI(
    title="Bell24h ML Service",
    description="SHAP/LIME explainability service for supplier matching and RFQ analysis",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Redis connection (using PostgreSQL as cache alternative)
import psycopg2
from psycopg2.extras import RealDictCursor

# Database connection for caching
def get_db_connection():
    return psycopg2.connect(
        host="ep-super-wind-a1c1ni4n-pooler.ap-southeast-1.aws.neon.tech",
        database="neondb",
        user="neondb_owner",
        password="npg_0Duqdxs3RoyA",
        port="5432",
        sslmode="require"
    )

# Cache management using PostgreSQL
class CacheManager:
    def __init__(self):
        self.conn = get_db_connection()
        self.create_cache_table()
    
    def create_cache_table(self):
        with self.conn.cursor() as cur:
            cur.execute("""
                CREATE TABLE IF NOT EXISTS ml_cache (
                    id SERIAL PRIMARY KEY,
                    cache_key VARCHAR(64) UNIQUE NOT NULL,
                    data JSONB NOT NULL,
                    created_at TIMESTAMP DEFAULT NOW(),
                    expires_at TIMESTAMP DEFAULT NOW() + INTERVAL '24 hours'
                );
                CREATE INDEX IF NOT EXISTS idx_cache_key ON ml_cache(cache_key);
                CREATE INDEX IF NOT EXISTS idx_expires_at ON ml_cache(expires_at);
            """)
            self.conn.commit()
    
    def get(self, key: str):
        with self.conn.cursor(cursor_factory=RealDictCursor) as cur:
            cur.execute(
                "SELECT data FROM ml_cache WHERE cache_key = %s AND expires_at > NOW()",
                (key,)
            )
            result = cur.fetchone()
            return json.loads(result['data']) if result else None
    
    def set(self, key: str, data: Any, ttl_hours: int = 24):
        with self.conn.cursor() as cur:
            cur.execute(
                "INSERT INTO ml_cache (cache_key, data, expires_at) VALUES (%s, %s, %s) ON CONFLICT (cache_key) DO UPDATE SET data = EXCLUDED.data, expires_at = EXCLUDED.expires_at",
                (key, json.dumps(data), datetime.now() + timedelta(hours=ttl_hours))
            )
            self.conn.commit()

cache = CacheManager()

# Lightweight ML models (pre-computed for efficiency)
class LightweightMLModels:
    def __init__(self):
        self.supplier_model = self._load_supplier_model()
        self.rfq_model = self._load_rfq_model()
        self.explainer_cache = {}
    
    def _load_supplier_model(self):
        # Create a lightweight model for demonstration
        # In production, load pre-trained model
        from sklearn.ensemble import RandomForestClassifier
        from sklearn.datasets import make_classification
        
        # Generate sample data for demo
        X, y = make_classification(n_samples=1000, n_features=10, random_state=42)
        model = RandomForestClassifier(n_estimators=10, random_state=42)
        model.fit(X, y)
        return model
    
    def _load_rfq_model(self):
        # Create a lightweight model for RFQ analysis
        from sklearn.ensemble import RandomForestRegressor
        from sklearn.datasets import make_regression
        
        X, y = make_regression(n_samples=1000, n_features=8, random_state=42)
        model = RandomForestRegressor(n_estimators=10, random_state=42)
        model.fit(X, y)
        return model

models = LightweightMLModels()

# Utility functions
def generate_cache_key(data: Dict[str, Any]) -> str:
    """Generate a unique cache key for the input data"""
    data_str = json.dumps(data, sort_keys=True)
    return hashlib.md5(data_str.encode()).hexdigest()

def extract_features(data: Dict[str, Any]) -> np.ndarray:
    """Extract features from input data"""
    # Map input data to feature vector
    features = []
    
    # Supplier matching features
    if 'supplier_data' in data:
        supplier = data['supplier_data']
        features.extend([
            supplier.get('experience_years', 0),
            supplier.get('rating', 0),
            supplier.get('response_time_hours', 24),
            supplier.get('completion_rate', 0),
            supplier.get('price_competitiveness', 0)
        ])
    
    # RFQ analysis features
    if 'rfq_data' in data:
        rfq = data['rfq_data']
        features.extend([
            rfq.get('budget', 0),
            rfq.get('urgency_score', 0),
            rfq.get('complexity_score', 0)
        ])
    
    # Pad or truncate to expected feature count
    expected_features = 10
    while len(features) < expected_features:
        features.append(0)
    
    return np.array(features[:expected_features]).reshape(1, -1)

# API Endpoints
@app.get("/")
async def root():
    return {
        "service": "Bell24h ML Service",
        "status": "running",
        "version": "1.0.0",
        "features": ["SHAP", "LIME", "Supplier Matching", "RFQ Analysis"]
    }

@app.get("/health")
async def health_check():
    return {
        "status": "healthy",
        "service": "ml-service",
        "memory_usage": "optimized",
        "shap_available": True,
        "lime_available": True,
        "cache_status": "connected",
        "models_loaded": True
    }

@app.post("/explain/supplier-matching")
async def explain_supplier_matching(data: Dict[str, Any]):
    """
    Explain supplier matching using SHAP values
    """
    try:
        # Generate cache key
        cache_key = f"supplier_matching_{generate_cache_key(data)}"
        
        # Check cache first
        cached_result = cache.get(cache_key)
        if cached_result:
            return cached_result
        
        # Extract features
        features = extract_features(data)
        
        # Get prediction
        prediction = models.supplier_model.predict(feature_features)[0]
        prediction_proba = models.supplier_model.predict_proba(features)[0]
        
        # Generate SHAP explanation (lightweight version)
        # In production, use actual SHAP library
        feature_names = [
            'experience_years', 'rating', 'response_time', 'completion_rate',
            'price_competitiveness', 'budget', 'urgency', 'complexity',
            'location_score', 'certification_score'
        ]
        
        # Simulate SHAP values (in production, use actual SHAP)
        shap_values = np.random.normal(0, 0.1, features.shape[1])
        shap_values = shap_values.tolist()
        
        # Create explanation
        explanation = {
            "prediction": int(prediction),
            "confidence": float(max(prediction_proba)),
            "shap_values": shap_values,
            "feature_names": feature_names,
            "feature_importance": dict(zip(feature_names, shap_values)),
            "explanation": "This supplier matches your requirements based on experience, rating, and response time.",
            "recommendations": [
                "High experience score indicates reliable supplier",
                "Good rating suggests quality service",
                "Fast response time shows good communication"
            ]
        }
        
        # Cache result
        cache.set(cache_key, explanation, ttl_hours=24)
        
        return explanation
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error generating explanation: {str(e)}")

@app.post("/explain/rfq-analysis")
async def explain_rfq_analysis(data: Dict[str, Any]):
    """
    Explain RFQ analysis using LIME
    """
    try:
        # Generate cache key
        cache_key = f"rfq_analysis_{generate_cache_key(data)}"
        
        # Check cache first
        cached_result = cache.get(cache_key)
        if cached_result:
            return cached_result
        
        # Extract features
        features = extract_features(data)
        
        # Get prediction
        prediction = models.rfq_model.predict(features)[0]
        
        # Generate LIME explanation (lightweight version)
        feature_names = [
            'budget', 'urgency', 'complexity', 'category_score',
            'supplier_count', 'market_demand', 'seasonality', 'location_factor'
        ]
        
        # Simulate LIME explanation
        lime_explanation = {
            "prediction": float(prediction),
            "feature_contributions": dict(zip(feature_names, np.random.normal(0, 0.1, len(feature_names)))),
            "explanation": "This RFQ has good market potential based on budget and urgency factors.",
            "confidence": 0.85,
            "recommendations": [
                "Consider increasing budget for better supplier response",
                "High urgency may limit supplier options",
                "Complexity score suggests detailed specifications needed"
            ]
        }
        
        # Cache result
        cache.set(cache_key, lime_explanation, ttl_hours=24)
        
        return lime_explanation
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error generating explanation: {str(e)}")

@app.post("/predict/supplier-score")
async def predict_supplier_score(data: Dict[str, Any]):
    """
    Predict supplier matching score
    """
    try:
        features = extract_features(data)
        prediction = models.supplier_model.predict_proba(features)[0]
        
        return {
            "supplier_score": float(max(prediction)),
            "confidence": float(max(prediction)),
            "recommendation": "Good match" if max(prediction) > 0.7 else "Consider alternatives"
        }
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error predicting supplier score: {str(e)}")

@app.post("/predict/rfq-success")
async def predict_rfq_success(data: Dict[str, Any]):
    """
    Predict RFQ success probability
    """
    try:
        features = extract_features(data)
        prediction = models.rfq_model.predict(features)[0]
        
        return {
            "success_probability": float(prediction),
            "confidence": 0.8,
            "recommendation": "High success probability" if prediction > 0.7 else "Consider revising requirements"
        }
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error predicting RFQ success: {str(e)}")

@app.get("/cache/stats")
async def cache_stats():
    """
    Get cache statistics
    """
    try:
        with cache.conn.cursor() as cur:
            cur.execute("SELECT COUNT(*) as total, COUNT(CASE WHEN expires_at > NOW() THEN 1 END) as active FROM ml_cache")
            result = cur.fetchone()
            
        return {
            "total_entries": result[0],
            "active_entries": result[1],
            "cache_hit_rate": "85%",  # Simulated
            "memory_usage": "optimized"
        }
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error getting cache stats: {str(e)}")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8001)
